Extracted some code to internals.go leaving geohex.go clean
